requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.3.0
scikit-learn>=0.24.2
pillow>=8.3.1
wandb>=0.12.0
PyYAML>=5.4.1
einops>=0.3.2
tqdm>=4.62.2
matplotlib>=3.4.3
seaborn>=0.11.2

